Create sampleandhold.cpp
First attempt at a plugin, based on the excellent tutorial by Anna Lantz: https://dev.epicgames.com/community/learning/tutorials/ry7p/unreal-engine-creating-metasound-nodes-in-c-quickstart

The first thing I wanted to try in Metasounds was a sample and hold, which I'm used to triggering at audio rate. I'm getting used to the native object now, but it made sense to try this as a first project.